cleaned up backend
