+++ controllers/auth.controller.js
@@ -74,11 +74,11 @@ const signout = (req, res) => {
 // has a valid JWT in the Authorization header
 // this function returns an error response
 // for an individual route
-// const verifyRequest = expressJwt({
-//   secret: config.jwtSecret,
-//   userProperty: "auth",
-//   algorithms: ["HS256"],
-// });
+const verifyRequest = expressJwt({
+  secret: config.jwtSecret,
+  userProperty: "auth",
+  algorithms: ["HS256"],
+});
 
 // check if user has auth to perfrom actions
 const hasAuthorization = (req, res, next) => {
@@ -95,5 +95,6 @@ const hasAuthorization = (req, res, next) => {
 module.exports = {
   signin,
   signout,
+  verifyRequest,
   hasAuthorization,
 };
