+++ models/user.model.js
@@ -24,14 +24,6 @@ const UserSchema = new mongoose.Schema({
     type: Date,
     default: Date.now,
   },
-  about: {
-    type: String,
-    trim: true,
-  },
-  photo: {
-    data: Buffer,
-    contentType: String,
-  },
 });
 
 UserSchema.virtual("password")
