--- routes/api/posts.routes.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-// @route GET api/posts
-// @desc  Test route
-// @access Public
-router.get('/', (req, res) => res.send('Posts route'));
-
-module.exports = router;
