--- routes/api/profile.routes.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-// @route GET api/profile
-// @desc  Test route
-// @access Public
-router.get('/', (req, res) => res.send('Profile route'));
-
-module.exports = router;
