+++ server.js
@@ -17,8 +17,7 @@ app.get("/", (req, res) => res.send("API Running"));
 // Define Routes
 app.use("/api/users", require("./routes/api/users.routes"));
 app.use("/api/auth", require("./routes/api/auth.routes"));
-app.use("/api/profile", require("./routes/api/profile.routes"));
-app.use("/api/posts", require("./routes/api/posts.routes"));
+app.use("/api/assets", require("./routes/api/assets.routes"))
 
 // specify port
 const PORT = process.env.PORT || 5000;
